refactor: CustomException 패키지 경로 변경
- CustomException 클래스의 패키지 경로 변경으로 인해 CustomException을 적용한 클래스 내에서도 해당 경로를 반영
